Implement wave access timing
When reading to Wave RAM, the byte read will be read from the current
sample position. On the DMG, this only happens if the APU read that byte
on the same cycle. (Similarly for writes.)

Since the APU period is 2 cycles, but an instruction is at least 4
cycles, it's possible that a sample was updated "in between" the
granularity of the instruction counter (which is driven by the
instruction timings). To work around this, the last 2 wave samples are
stored with their access time.

This change makes dmg_sound-2 tests 09 and 12 pass.