Implement wave channel trigger corruption bug
Triggering the wave channel while it is playing, and it has just read
from wave ram, will corrupt wave ram. This is tested by dmg_sound-2 test
10.

Not super happy with the weird offsets I need to make this pass, but
it's not a huge hack, I guess.

Now all of dmg_sound-2 tests pass!