Fix sprite priority
The document I read originally made it seem as though the visible
sprites (if more than 10 per on a scanline) were the ones that had
smaller X-coordinates. Nope, the visible sprites are the ones that occur
earliest in the OAM table, but the ones with smaller X-coordinates are
drawn on top.