Fix stat_write_if-GS
STAT should be triggered if LY=LYC, and its IRQ is not currently set.
It's not clear from the test whether this would work when writing STAT
if the new value for the LY=LYC IRQ is false, but as this is not
necessary for the hblank and vblank IRQs, I assume it's not necessary
here either.