Properly implement HALT bug
I didn't realize there was another Blargg test for this, thanks
DParrott!

Basically, the behavior is this:

IME=1: enter HALT mode, call interrupt vector on interrupt
IME=0, IE & IF & 0x1f == 0: enter HALT mode, don't call interrupt vector
IME=0, IE & IF & 0x1f != 0: HALT bug

This kinda makes sense, now that I think of it. Basically, the HALT bug
is normal behavior that you'd expect if IME=0, but the interrupt is
firing immediately. This screws up the opcode fetch somehow so the
opcode is fetched at the correct location, but the opcode parameters are
not.