Reading from non-existant ext ram should yield FF
This fixes Thunderbirds:

Thunderbirds sets two interrupts, VBlank and STAT at LY=0. Address
0xC0A1 is set to 1 when the VBlank interrupt fires, and set to 0 when
STAT interrupt fires.

After the title screen fades, VBlank fires, then the interrupt routine
runs long enough that the STAT interrupt fires too. This sets 0xC0A1 to
0. Then the screen is disabled and IME is set to 0.

The game then starts running a HALT loop, waiting for 0xC0A1 to be
non-zero. This won't happen, since not only is the screen disabled (so
no LCD interrupts will fire), but also IME=0.

Except! since HALT is executed with IME=0 and IF & IE != 0, this causes
the "halt bug". This turns:

  HALT
  LD A, (0xC0A1)

into

  HALT
  LD A, (0xA1FA)
  RET NZ

The memory at 0xA1FA maps to external RAM, but this cartridge doesn't
have any. This should return 0xFF (the value that is always returned for
unmapped data), but was returning 0 instead, which caused an infinite
loop.