Fix intr2_mode0_timing_sprites
This formula for sprite timing works on the NOPS tests too, but those
have an additional 1 M-cycle delay. I'm guessing this comes from the
HALT instruction. If I add in that delay, a number of other tests break.

I think the solution is to adjust the timing of when IF gets set on an
HBLANK trigger, but I haven't tried it yet.