Add support for MBC1M and MMM01 mappers
We scan the ROM for any location were we can find the logo. This seems
to be a strong enough heuristic for all MBC1M and MMM01 games I tried.

The MMM01 header should always be at the end in a properly dumped ROM.
MBC1M works like a normal MBC1, but can swap the both ROM banks to a new
location.

The implementation of MMM01 is just enough to work, but it's hard to
know what the values written to ROM are supposed to mean because there
are only two official games that use it.

For MBC1M, the behavior is discussed more here:
http://forums.nesdev.com/viewtopic.php?f=20&t=14099. This implementation
is much stupider, and assumes that you only ever switch away from
multi-cart mode. This seems to match the behavior in my testing.