Fix emscripten build
Not sure why, but I switched init_audio_buffer to taking an AudioBuffer
instead of an Emulator. This broke the emscripten build, which still was
passing an Emulator to the function (not typechecked).

It's also necessary to switch the cycle-count loop in renderVideo to use
a 32-bit unsigned value. Otherwise runUntilCycles will use a JS number,
which can be > 2**32, then compare it to the result of _get_cycles,
which can't.