Synchronize video to the monitor refresh rate
This simplifies a lot of the host code; we can now just rely on the
vsync to synchronize, rather than many calls to `SDL_Delay`.

I've also changed `emulator_run` again to `emulator_run_until`, which
runs until the emulator reaches a given cycle count. This simplifies a
bunch of other code too; now the tester can just use this, as well as
the emscripten code. This did break the timing for a number of tests,
many of which needed to run for a few more frames.

I also added `host_run_ms` which runs is used to run one frame's worth
of time. This makes it easier to share code between `binjgb.c` and
`debugger.cc`.

Finally, I separated out `host_upload_video` from `host_render_video`.
This uploads the texture, but doesn't swap buffers.