OBJ window
The OBJ window has a 10x4 grid that displays the currently visible
sprites. It allows you to highlight the selected sprite in the display,
as well as displaying the basic OAM info (X, Y, tile index, X or Y flip,
etc.)

Loading a texture from tile data was also abstracted into the TileImage
class. The OBJ window has its own copy of the data so it can display
with the correct palette. It would be cool to use the same tile data and
use a shader to choose the correct palette instead, but this was easier.