Fix build on AppVeyor
Somehow ERROR gets defined (probably windows.h), so #undef it.

glBlendEquation and glActiveTexture are defined in SDL_opengl.h, and
have prototypes (with GLAPI and GLAPIENTRY) defined as well. But the
Windows GL library doesn't seem to implement these because they're not
in OpenGL 1.1. Fortunately, we never change the default values anyway,
so just remove the calls.