Use palette shader to avoid duplicate textures
The texture data sent down is only 8bpp (techically only 2bpp), so it's
nicer to just set uniforms in the shader to convert from 2bpp -> 32bpp.
This way we save having to upload the same texture with different
colors. It also makes it easier in the future to allow choosing the RGBA
values to represent the different DMG shades.