Handle `SDL_WINDOWEVENT_SHOWN` message
It doesn't seem as though a `SIZE_CHANGED` or `RESIZED` event is sent on
startup on Windows, so the display is black. This change listens for a
SHOWN message and does the same behavior.

Also use `SDL_WINDOWEVENT_SIZE_CHANGED` instead of
`SDL_WINDOWEVENT_RESIZED`, since `SIZE_CHANGED` is sent whenever the
size changes regardless of how it was done. `RESIZED` is only sent when
the size is changed externally (via the user or the system).

This fixes issue #11.