More precision w/ debug rewinding
Use two separate sliders, one for frames, one for cycles within that
frame. Also add buttons to increment/decrement.

Cycle seeking doesn't do this quite right, since it will always
overshoot if the cycle requested is in the middle of an instruction, but
it seems to work well enough for now.