Fix voice in Bart's escape from camp deadly
The audio is played through the noise channel by manually updating the
master volume. This was broken because the noise channel wasn't being
updated when its clock_shift value was > 13. It's correct to not update,
but it should still play its sample. The sample value must be
initialized to 1 as well for this to work too -- the LFSR was already
set properly.

Thanks to mattcurrie for the help!