Wrap display of ROM window data
Rather than displaying the ROM usage map scaled, it's much nicer to
display as much as possible per line and wrap so there is no horizontal
scrollbar.

I've also changed the "unused" color to a dark gray, so you can tell the
difference between that and memory that is out-of-bounds.