Rewind steps forward/backward exactly 1 instr.
The previous solution stepped forward/backward 10 cycles, but that's not
useful because many instructions are shorter or longer than that. So
that means that when you step forward, it may jump over instructions,
and when you step back it may not move back at all.

The solution here is cheesy, but works: rewind to 28 cycles before now,
save state, step forward until "now", then load state and step forward
again, but stop 1 step earlier.