Bunch of work on the Disassembly window
The primary difference here is how the disassembly is done. Every frame,
we iterate over the entire address space and disassemble everything.
This gives a fixed number of lines so we can display a proper scrollbar.

The ROM usage is also used to better find code. To do this, we track the
opcode "start", in addition to whether it is code vs. data.

Some other improvements:

* Colored register values
* Disassembly window is on top at startup
* Compress register display to fewer lines
* Add "goto" box

Also a bug fix:

* Use read_u8_raw when disassembling. Without this, disassembling will
  modify ROM usage.