Properly set the initial state of mbc5
ROM1 is set to bank 1 by default in mbc5. This was properly being done,
but if there was a write to $3000-3fff range (which affects the 8th bit
of the bank), the default value of the mbc5 state will be used, which
was 0. This was setting the bank to 0 instead of 1.

This fixes Fort Boyard.