Implement basic breakpoint support
Breakpoints and illegal opcodes always exit the run loop. Each tool will
handle it differently.

* tester will ignore breakpoints (they should never be set) but will
  display an error on illegal opcodes.
* binjgb is similar, though it will just pause on an illegal opcde, and
  allow you to continue.
* debugger will pause, but also has support for adding breakpoints.

Breakpoints are rendered as red circles on the left of the instruction.
Currently, it toggles between 3-states: enabled, disabled, and removed.

Checking for breakpoints only happens when the opcode is read. The
emulator_step hook was originally the place to do this (along with
marking ROM usage), but it's actually not a very good place because it
will be called even during halts and DMAs. Instead, I've moved the check
to a new hook that only occurs when the opcode is read. This means the
event has to bubble up through more layers, but the code is much faster.

The implementation is mostly straightforward, but has one small
optimization. All enabled breakpoints will have their addresses combined
to produce a set of required 0 and 1 bits. Then the PC can check against
this pattern to determine whether any breakpoint could possibly match.
If so, it linearly scans through all enabled breakpoints for the match.
Not sure how much this actually helps, though.

Using breakpoints with rewind "works", though the behavior is a bit
odd: it will first load the closest save state, then run forward to the
desired tick. If it hits a breakpoint, it will stop there.

TODO:
- add a breakpoint window.
- add temporary breakpoints, for "run to here".
- better rewind support (continue backward?)