Use Vue.js for UI stuff
I don't really know anything about JS frameworks, and I didn't want to
get bogged down in research (done that many times before). But I took a
look at Vue.js and it seems pretty simple to use and didn't require a
bunch of npm junk.

Not sure I'm using it right yet, but so far it's cleaned up the code
some. It'll be nice to have for future UI work as well.