Remove BREAKPOINT hack
Previously we communicated that a breakpoint was hit by returning a
special opcode. Instead, just set EmulatorState's event field.

Also move the breakpoint check back to emulator_step. It can't be during
opcode reading since we've already run `tick` by that point, so the next
time we step it will double-tick and desync.