Use ImGui docking branch, remove imgui_dock
The new docking system looks a bit nicer, is more stable, and persists
the docking layout to the imgui.ini file. I haven't seen any issues with
it yet.

The public layout API is not set yet, so I have to use the internal API
to set an initial layout.