Fix back-to-color demo (green pentagon)
The green pentagon is OBJ, and the grey pentagon is BG. The BG attribute
is normal, and the OBJ attribute is behind BG, so the green pentagon
would normally draw behind the gray one.

But on the CGB, if bit 0 is cleared then the OBJ always draws in front
of BG/window, regardless of any other attributes.