Fix crashes when rewinding and changing palette
The BWPalette structure is stored in the EmulatorState, which is saved
during rewind. I wanted to see the palette updated when you are
scrubbing through the rewind timeline, so I updated the BWPalette's
cachced RGBA values. Doing so can corrupt the rewind state, since some
frames are recreated using deltas from the current state.

The fix is to store the cached palettes in the Emulator, not the
EmulatorState. Now the palettes can be updated at any time without
affecting the rewind state.