Only synchronize when TICKS > sync_ticks
This fixes a hang in the debugger when reading I/O values during the
Donkey Kong "Mario" demo. The timer would attempt to synchronize when
TICKS < sync_ticks, which would try to synchronize with a huge
delta_ticks value.

It's arguable that the debugger shouldn't be able to synchronize the
subsystems anyway, but in general it should be safe. Synchronization
should only be an optimization and not affect the behavior regardless
how many times it's called.