Only read joypad from SDL polling loop
We used to read the current key state directly inside the joypad
callback. This has the potential benefit of providing better resolution
on keypresses, but has the drawback that it makes it annoying to use the
enter key in ImGui in an `InputText` box.

ImGui sets the `WantCaptureKeyboard` flag when it wants to prevent
keypresses from being sent to the main application. This happens
properly, but after enter is pressed the `InputText` box is no longer
active. The next frame, the `WantCaptureKeyboard` is no longer set, and
the keyboard callback is called, which will read the current state of
the enter key. Since the key is down, it will tell the emulator that the
start button is pressed.