Fix popping in wasm demo; add volume slider
The emulator always generates u8 samples [0,255], but both the SDL2 host
and the web host now use f32 samples [-1,1]. If we convert these to use
the full range, then 0 becomes -1 and 255 becomes 1. But when the audio
system is initialized it generates silence at 0, not -1. When the
emulator starts up it is generally silent as well, but will produce 0,
which is then converted to -1 and creates a pop.

A simple solution is to just use half the f32 range, [0,1]. It seems
that the SDL2 host was already doing this, so I converted the web host
to do the same.

I also added a simple volume slider, and removed the ROM name.